+++ app/resources/views/pages/dominion/realm.blade.php
@@ -35,7 +35,7 @@
                             </tr>
                         </thead>
                         <tbody>
-                            @for ($i = 0; $i < $round->realm_size; $i++)
+                            @for ($i = 0; $i < max($round->realm_size, $dominions->count()); $i++)
                                 @php
                                     $dominion = $dominions->get($i);
                                 @endphp
@@ -141,7 +141,7 @@
                             <div class="col-xs-4">
                                 <form action="{{ route('dominion.realm.change-realm') }}" method="post" role="form">
                                     @csrf
-                                    <input type="number" name="realm" class="form-control text-center" placeholder="{{ $realm->number }}" min="1" max="{{ $realmCount }}">
+                                    <input type="number" name="realm" class="form-control text-center" placeholder="{{ $realm->number }}" min="0" max="{{ $realmCount }}">
                                 </form>
                             </div>
                             <div class="col-xs-4 text-right">